resumo_video/requirements.txt
streamlit
openai>=1.0.0
python-dotenv
yt_dlp
pydub
openai-whisper
python-dotenv
ffmpeg-python